Move SentryProperties to sentry-core.
